#3766 move path to files remove map layer
