+++ src/main/java/edu/harvard/iq/dataverse/api/Files.java
@@ -5,14 +5,12 @@ import com.google.gson.JsonObject;
 import edu.harvard.iq.dataverse.DataFile;
 import edu.harvard.iq.dataverse.DataFileServiceBean;
 import edu.harvard.iq.dataverse.DatasetServiceBean;
-import edu.harvard.iq.dataverse.DatasetVersion;
 import edu.harvard.iq.dataverse.DatasetVersionServiceBean;
 import edu.harvard.iq.dataverse.DataverseRequestServiceBean;
 import edu.harvard.iq.dataverse.DataverseServiceBean;
 import edu.harvard.iq.dataverse.EjbDataverseEngine;
-import edu.harvard.iq.dataverse.MapLayerMetadata;
 import edu.harvard.iq.dataverse.UserNotificationServiceBean;
-import edu.harvard.iq.dataverse.authorization.Permission;
+import static edu.harvard.iq.dataverse.api.AbstractApiBean.error;
 import edu.harvard.iq.dataverse.authorization.users.User;
 import edu.harvard.iq.dataverse.datasetutility.AddReplaceFileHelper;
 import edu.harvard.iq.dataverse.datasetutility.DataFileTagException;
@@ -20,17 +18,13 @@ import edu.harvard.iq.dataverse.datasetutility.NoFilesException;
 import edu.harvard.iq.dataverse.datasetutility.OptionalFileParams;
 import edu.harvard.iq.dataverse.engine.command.DataverseRequest;
 import edu.harvard.iq.dataverse.engine.command.exception.CommandException;
-import edu.harvard.iq.dataverse.engine.command.impl.CreateDatasetCommand;
-import edu.harvard.iq.dataverse.engine.command.impl.CreateDatasetVersionCommand;
 import edu.harvard.iq.dataverse.engine.command.impl.DeleteMapLayerMetadataCommand;
 import edu.harvard.iq.dataverse.engine.command.impl.RestrictFileCommand;
+import edu.harvard.iq.dataverse.engine.command.impl.UningestFileCommand;
 import edu.harvard.iq.dataverse.engine.command.impl.UpdateDatasetCommand;
-import edu.harvard.iq.dataverse.engine.command.impl.UpdateDatasetVersionCommand;
 import edu.harvard.iq.dataverse.ingest.IngestServiceBean;
 import edu.harvard.iq.dataverse.util.SystemConfig;
-import static edu.harvard.iq.dataverse.util.json.JsonPrinter.json;
 import java.io.InputStream;
-import java.sql.Timestamp;
 import java.util.ResourceBundle;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -274,7 +268,7 @@ public class Files extends AbstractApiBean {
         }
             
     } // end: replaceFileInDataset
-
+    
     @DELETE
     @Path("{id}/map")
     public Response getMapLayerMetadatas(@PathParam("id") Long idSupplied) {
@@ -295,6 +289,31 @@ public class Files extends AbstractApiBean {
         } catch (CommandException ex) {
             return error(BAD_REQUEST, "Problem trying to delete map from file id " + dataFile.getId() + ": " + ex.getLocalizedMessage());
         }
+    }
+    
+    @Path("{id}/uningest")
+    @POST
+    public Response uningestDatafile(@PathParam("id") Long idSupplied) {
+
+        DataFile dataFile = fileService.find(idSupplied);
+
+        if (dataFile == null) {
+            return error(Response.Status.NOT_FOUND, "File not found for given id.");
+        }
+
+        if (!dataFile.isTabularData()) {
+            return error(Response.Status.BAD_REQUEST, "Cannot uningest non-tabular file.");
+        }
+
+        try {
+            DataverseRequest req = createDataverseRequest(findUserOrDie());
+            execCommand(new UningestFileCommand(req, dataFile));
+            return ok("Datafile " + idSupplied + " uningested.");
+
+        } catch (WrappedResponse wr) {
+            return wr.getResponse();
+        }
+
     }
 
 }
